requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.15
django-cors-headers==4.4.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-storages==1.14.4
django-upgrade==1.19.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
gevent==24.2.1
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai==0.7.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
loguru==0.7.2
lxml==5.3.0
MarkupSafe==2.1.5
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pillow==10.3.0
playwright==1.47.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
scrapfly-sdk==0.8.18
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
snscrape==0.7.0.20230622
socksio==1.0.0
soupsieve==2.6
sqlparse==0.5.0
tokenize-rt==6.0.0
tqdm==4.66.5
tweepy==4.14.0
twikit==2.1.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.7.0
win32-setctime==1.1.0
wrapt==1.16.0
zope.event==5.0
zope.interface==7.0.3
